Determin shared accounts to from reusable selector
