+++ app/core/selectors.js
@@ -1,0 +1,65 @@
+import { filter, get, has, isEmpty, map, reject, values } from 'lodash';
+import personUtils from './personutils';
+import { createSelector } from 'reselect';
+
+export const selectPatientSharedAccounts = createSelector([state => state.blip], state => {
+  const {
+    allUsers,
+    clinics,
+    loggedInUserId,
+    membersOfTargetCareTeam,
+    pendingSentInvites,
+    permissionsOfMembersInTargetCareTeam,
+  } = state;
+
+  const pendingInvites = reject(pendingSentInvites, personUtils.isDataDonationAccount);
+  const pendingMemberInvites = filter(pendingInvites, ({ email }) => !isEmpty(email));
+  const patientClinics = filter(values(clinics), ({ patients }) => has(patients, loggedInUserId));
+  const clinicInvites = filter(pendingInvites, ({ clinicId }) => !isEmpty(clinicId));
+
+  const accounts = [
+    ...(map(patientClinics, clinic => ({
+      id: clinic.id,
+      name: clinic.name,
+      nameOrderable: clinic.name.toLowerCase(),
+      permissions: get(clinic, ['patients', loggedInUserId, 'permissions']),
+      role: 'clinic',
+      type: 'clinic',
+      uploadPermission: !!get(clinic, ['patients', loggedInUserId, 'permissions', 'upload']),
+    }))),
+    ...(map(membersOfTargetCareTeam, memberId => ({
+      email: get(allUsers, [memberId, 'emails', '0']),
+      id: get(allUsers, [memberId, 'userid']),
+      name: personUtils.fullName(allUsers[memberId]),
+      nameOrderable: (personUtils.fullName(allUsers[memberId]) || '').toLowerCase(),
+      permissions: get(permissionsOfMembersInTargetCareTeam, [memberId]),
+      role: personUtils.hasClinicRole(allUsers[memberId]) ? 'clinician' : 'member',
+      type: 'account',
+      uploadPermission: !!get(permissionsOfMembersInTargetCareTeam, [memberId, 'upload']),
+    }))),
+    ...(map(pendingMemberInvites, invite => ({
+      email: invite.email,
+      key: invite.key,
+      nameOrderable: invite.email,
+      permissions: invite.context,
+      role: 'member',
+      status: invite.status,
+      type: invite.type,
+      uploadPermission: !!get(invite, ['context', 'upload']),
+      created: invite.created,
+    }))),
+    ...(map(clinicInvites, invite => ({
+      id: invite.clinicId,
+      key: invite.key,
+      name: get(clinics, [invite.clinicId, 'name'], ''),
+      nameOrderable: get(clinics, [invite.clinicId, 'name'], '').toLowerCase(),
+      permissions: invite.context,
+      role: 'clinic',
+      status: invite.status,
+      type: invite.type,
+      uploadPermission: !!get(invite, ['context', 'upload']),
+    }))),
+  ];
+
+  return accounts;
+});
